// src/api/parameters/getNotificationsIdField.ts
export interface GetNotificationsIdField {
  /** The API key to use */
  key: Record<string, any>;
  /** The API token to use */
  token: Record<string, any>;
  /** The ID of the notification */
  id: Record<string, any>;
  /** A notification [field](/cloud/trello/guides/rest-api/object-definitions/) */
  field: Record<string, any>;
}